Removed IListenerBase (unused) and extracted its PendingTransactionState to a separate class
