+++ modAionImpl/src/org/aion/zero/impl/blockchain/AionImpl.java
@@ -3,7 +3,6 @@ package org.aion.zero.impl.blockchain;
 import java.util.List;
 import java.util.Optional;
 import org.aion.base.AionTransaction;
-import org.aion.base.TxUtil;
 import org.aion.crypto.ECKey;
 import org.aion.crypto.ECKeyFac;
 import org.aion.equihash.EquihashMiner;
@@ -220,7 +219,7 @@ public class AionImpl implements IAionChain {
         return aionHub;
     }
 
-    private IPendingStateInternal<?> getIPendingStateInternal() {
+    private IPendingStateInternal getIPendingStateInternal() {
         return this.aionHub.getPendingState();
     }
