+++ modMcf/src/org/aion/mcf/blockchain/IPendingStateInternal.java
@@ -1,17 +1,9 @@
 package org.aion.mcf.blockchain;
 
-import java.util.List;
-import org.aion.mcf.types.AbstractTxReceipt;
-
 /**
  * Internal pending state interface.
- *
- * @param <BLK>
  */
-public interface IPendingStateInternal<BLK extends Block> extends IPendingState {
-
-    // called by onBest
-    void processBest(BLK block, List<? extends AbstractTxReceipt> receipts);
+public interface IPendingStateInternal extends IPendingState {
 
     void shutDown();
